% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{prep data for the testing functions}
\usage{
prep_data(gene_data, feature_num = Inf, label_vector = "ensg00000183117_19")
}
\arguments{
\item{gene_data}{a data.frame where the label_vector column is the response
and the rest of the columns are predictors}

\item{feature_num}{is the number of feature_num to select from the data. must be
$>=$ 1.If exceeds ncol, ncol-1 is used (1 col removed as the label).
Default is to use all columns}

\item{label_vector}{name of the column to use as the response}
}
\value{
a list with xgboost data matrix objects, slots dtrain, dtest and wl
see the xgboost docs on wl. dtrain is the trainin data to
}
\description{
prep data for the testing functions
}
\examples{
  test_data = readRDS(
      system.file('testing_gene_data.rds',
      package = 'brentlabModelPerfTesting'))
  # note: suppressed warnings used here b/c the test data is too
  # small to stratify. Typically, do not use supressWarnings.
  suppressWarnings({prepped_data_subset = brentlabModelPerfTesting::prep_data(test_data, 10)})

  names(prepped_data_subset)

}
